requirements.txt
Django>=3.0,<4.0
# psycopg2>=2.8
django-service-objects==0.7.0
factory_boy==3.2.1
